requirements.txt
pypdf
langchain
chromadb
pytest
boto3
langchain-community
langchain-chroma
pyyaml
tqdm
